Refactored `gc_fasta_to_orf_seq_dict` for simplicity.
